Add files via upload
The varying p has a rather fixed value of p, it solves the acceptance probability for values of p ranging from 0.1 to 0.9. the value can be altered by changing the range in the code

For the other part, the user can easily enter the value of p for which they are interested and other parameters.